civicpulse-backend/requirements.txt
fastapi
uvicorn
boto3
langchain
opensearch-py
pydantic
python-multipart
tiktoken
unstructured
requests
beautifulsoup4
python-jose[cryptography]
python-dotenv
openai
elevenlabs
